:sparkles: Pull specifications locally
Prior this commit, the operator is relying on gravitee to make a request to
the candidate service. This causes an issue when the service resides in a
different namespace from gravitee. To ensure that the OAS is accessible, the
operator now requests for the OAS locally, before sending it to gravitee for
import.